+++ Byte.Core.Models/RoleDTO.cs
@@ -19,7 +19,7 @@ namespace Byte.Core.Models
         /// <summary>
         /// 菜单
         /// </summary>
-        public int[]? MenuIds { get; set; }
+        public long[]? MenuIds { get; set; }
     }
 
     /// <summary>
@@ -50,7 +50,7 @@ namespace Byte.Core.Models
         /// <summary>
         /// 主键Id!
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// 账号
@@ -66,7 +66,7 @@ namespace Byte.Core.Models
     public class RoleInfo : Role
     {
 
-        public int[] MenuIds { get; set; }
+        public long[] MenuIds { get; set; }
 
     }
 }
