+++ Byte.Core.Repository/DeptRepository.cs
@@ -6,7 +6,7 @@ namespace Byte.Core.Repository
     /// <summary>
     /// 部门
     /// </summary>
-    public class DeptRepository : BaseRepository<int, Dept>
+    public class DeptRepository : BaseRepository<long, Dept>
     {
         public DeptRepository(IUnitOfWork unitOfWork) : base(unitOfWork)
         {
