+++ Byte.Core.Repository/Role_MenuRepository.cs
@@ -6,7 +6,7 @@ namespace Byte.Core.Repository
     /// <summary>
     /// 角色-菜单
     /// </summary>
-    public class Role_MenuRepository : BaseRepository<int, Role_Menu>
+    public class Role_MenuRepository : BaseRepository<long, Role_Menu>
     {
         public Role_MenuRepository(IUnitOfWork unitOfWork) : base(unitOfWork)
         {
