+++ Byte.Core.Tools/Attributes/JWTPayload.cs
@@ -10,7 +10,7 @@ namespace Byte.Core.Tools.Attributes
         /// <summary>
         /// 用户的唯一标识符
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// 用户的账户
@@ -30,7 +30,7 @@ namespace Byte.Core.Tools.Attributes
         /// <summary>
         /// 用户所属部门的唯一标识符
         /// </summary>
-        public int DeptId { get; set; }
+        public long DeptId { get; set; }
 
         /// <summary>
         /// 用户的角色代码
